pyproject.toml
[tool.poetry]
name = "wire-service"
version = "0.1.0"
description = ""
authors = ["Maik Schöpe <m.schoepe@ccgmbh.de>"]
readme = "README.md"
packages = [{include = "wire_service"}]

[tool.poetry.dependencies]
python = "^3.8"
dynaconf = "^3.1.7"
psycopg2-binary = "^2.9.3"
strawberry-graphql = {extras = ["asgi"], version = "^0.103.9"}
SQLAlchemy = {extras = ["mypy"], version = "^1.4.32"}
uvicorn = {extras = ["standard"], version = "^0.17.6"}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
mypy = "^0.940"
flake8 = "^4.0.1"
alembic = "^1.7.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin,strawberry.ext.mypy_plugin"
files = "wire_service,alembic"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
]